+++ 1.20/Common/src/main/java/fuzs/puzzleslib/impl/client/core/context/BlockRenderTypesContextImpl.java
@@ -3,13 +3,12 @@ package fuzs.puzzleslib.impl.client.core.context;
 import com.google.common.base.Preconditions;
 import fuzs.puzzleslib.api.client.core.v1.ClientAbstractions;
 import fuzs.puzzleslib.api.client.core.v1.context.RenderTypesContext;
-import net.fabricmc.fabric.api.blockrenderlayer.v1.BlockRenderLayerMap;
 import net.minecraft.client.renderer.RenderType;
 import net.minecraft.world.level.block.Block;
 
 import java.util.Objects;
 
-public final class BlockRenderTypesContextFabricImpl implements RenderTypesContext<Block> {
+public final class BlockRenderTypesContextImpl implements RenderTypesContext<Block> {
 
     @Override
     public void registerRenderType(RenderType renderType, Block... blocks) {
@@ -18,7 +17,7 @@ public final class BlockRenderTypesContextFabricImpl implements RenderTypesConte
         Preconditions.checkPositionIndex(1, blocks.length, "blocks is empty");
         for (Block block : blocks) {
             Objects.requireNonNull(block, "block is null");
-            BlockRenderLayerMap.INSTANCE.putBlock(block, renderType);
+            ClientAbstractions.INSTANCE.registerRenderType(block, renderType);
         }
     }